requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Deprecated==1.2.14
idna==3.7
jwt==1.3.1
pycparser==2.22
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.31.0
typing_extensions==4.11.0
urllib3==2.2.1
wrapt==1.16.0
